+++ crates/amaru-node/src/tests/test_data.rs
@@ -21,13 +21,13 @@ use amaru_kernel::{
 };
 use amaru_ouroboros::{Mempool, TxInsertResult, TxOrigin};
 
-pub fn create_transactions(number: usize) -> Vec<WithOriginalBytes<Transaction>> {
+pub fn create_transactions(number: u16) -> Vec<WithOriginalBytes<Transaction>> {
     (0..number).map(create_transaction).collect()
 }
 
 pub fn create_transactions_in_mempool(
     mempool: Arc<dyn Mempool<WithOriginalBytes<Transaction>>>,
-    number: usize,
+    number: u16,
 ) -> Vec<WithOriginalBytes<Transaction>> {
     let mut txs = vec![];
     for i in 0..number {
@@ -40,8 +40,8 @@ pub fn create_transactions_in_mempool(
 }
 
 /// Create a transaction with a unique input based on the given id.
-pub fn create_transaction(id: usize) -> WithOriginalBytes<Transaction> {
-    let tx_input = TransactionInput { transaction_id: Hash::new([1; TRANSACTION_BODY]), index: id as u64 };
+pub fn create_transaction(id: u16) -> WithOriginalBytes<Transaction> {
+    let tx_input = TransactionInput { transaction_id: Hash::new([1; TRANSACTION_BODY]), index: id };
 
     let body = TransactionBody::new([tx_input], [], 0);